Move all auth-related stuff to AuthenticationHelper
